Added custom WAV file reader.
With javax.sound.sampled there is no way to read an audio file without
probing for the file format. When the file format is known in advance,
testing all known formats is not necessary. This was done to circumvent
a timing problem when establishing RTP connections to Fritz!Box routers.
The hope was to get a faster startup without file format probing. When
the first RTP packet has a delay of 200ms, the audio connection is not
established. However, avoiding file format probing did not lower the
startup time sufficiently.