Bugfix: Use contact address type corresponding to route header.
This ensures that a IPv6 address is used as contact address, if a client
is registered on a SIP server with an IPv6 address. Since the SIP server
contacts the client via its contact address, address types must match.

However, if a SIP server has an IPv4 and IPv6 address, registration
works, even if address types differ. When registering with an IPv4
address to an IPv6 SIP server that also has an IPv4 address, an INVITE
is sent to the IPv4 address of the client with a VIA header pointing to
the IPv6 address of the server. While this should work in principal, the
connection attempt seems to fail randomly (?) in practice with Fritz!Box
SIP servers. Making contact address types match the route address type
is an experiment making connection setup more reliable. In general,
selecting a contact address type according to the route address, should
make things at least not worse.